+++ lib/appbar.dart
@@ -5,7 +5,7 @@ import 'package:money/models/constants.dart';
 import 'package:money/models/settings.dart';
 import 'package:money/storage/import/import_transactions_from_text.dart';
 import 'package:money/views/view_currencies.dart';
-import 'package:money/widgets/change_summary.dart';
+import 'package:money/views/view_pending_changes/bage_pending_changes.dart';
 import 'package:money/widgets/three_part_label.dart';
 import 'package:money/widgets/zoom.dart';
 
@@ -222,7 +222,7 @@ class _MyAppBarState extends State<MyAppBar> {
                 const SizedBox(
                   width: 8,
                 ),
-                ChangeSummaryBadge(
+                BadgePendingChanges(
                   itemsAdded: Settings().trackMutations.added,
                   itemsChanged: Settings().trackMutations.changed,
                   itemsDeleted: Settings().trackMutations.deleted,
@@ -239,7 +239,7 @@ class _MyAppBarState extends State<MyAppBar> {
                 const SizedBox(
                   width: 8,
                 ),
-                ChangeSummaryBadge(
+                BadgePendingChanges(
                   itemsAdded: Settings().trackMutations.added,
                   itemsChanged: Settings().trackMutations.changed,
                   itemsDeleted: Settings().trackMutations.deleted,