+++ lib/models/money_objects/accounts/accounts.dart
@@ -7,6 +7,10 @@ import 'package:money/models/money_objects/money_objects.dart';
 import 'package:money/models/money_objects/transactions/transaction.dart';
 
 class Accounts extends MoneyObjects<Account> {
+  Accounts() {
+    collectionName = 'Accounts';
+  }
+
   @override
   Account instanceFromSqlite(final MyJson row) {
     return Account.fromJson(row);