+++ lib/models/money_objects/online_accounts/online_accounts.dart
@@ -3,6 +3,10 @@ import 'package:money/models/money_objects/money_objects.dart';
 import 'package:money/models/money_objects/online_accounts/online_account.dart';
 
 class OnlineAccounts extends MoneyObjects<OnlineAccount> {
+  OnlineAccounts() {
+    collectionName = 'Online Accounts';
+  }
+
   @override
   void loadFromJson(final List<MyJson> rows) {
     clear();